kfs: arch: build.zig support for multiarch (wip)
empty stubs for x86_64 arch
replacing of u32 to usize (wip)

+++ src/kernel/fs/fs.zig
@@ -46,7 +46,7 @@ pub const procfs = @import("proc/filesystem.zig");
 
 const std = @import("std");
 pub const DentryHash = struct {
-    sb: u32,
+    sb: usize,
     ino: u32,
     name: []const u8,
 };
